requirements.txt
aniso8601==8.0.0
asgiref==3.2.7
Django==3.1a1
django-graphql-extensions==0.0.7
django-graphql-jwt==0.3.1
django-seed==0.2.2
Faker==2.0.5
graphene==3.0b2
graphene-django==3.0.0b2
graphql-core==3.1.1
graphql-relay==3.0.0
Pillow==7.1.2
promise==2.3
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
Rx==1.6.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
Unidecode==1.1.1
